Fix binary compatible problems
