+++ slick/src/main/scala-2.13+/slick/util/ConstArrayCompat.scala
@@ -1,0 +1,13 @@
+package slick.util
+
+trait ConstArrayCompat {
+  def from[T](values: scala.collection.immutable.Iterable[T]): ConstArray[T] = {
+    val a = new Array[Any](values.size)
+    var i = 0
+    values.foreach { v =>
+      a(i) = v
+      i += 1
+    }
+    new ConstArray[T](a)
+  }
+}
