+++ slick/src/main/scala-2.13-/slick/util/SQLBuilderCompat.scala
@@ -1,0 +1,14 @@
+package slick.util
+
+trait SQLBuilderCompat {
+
+  def +=(s: String): SQLBuilder
+
+  def sep[T](sequence: scala.collection.Traversable[T], separator: String)(f: T => Unit): Unit = {
+    var first = true
+    for(x <- sequence) {
+      if(first) first = false else this += separator
+      f(x)
+    }
+  }
+}
